refactor: change report generator component location
